requirements.txt
fastapi
uvicorn
pyttsx3
python-dotenv
simpleaudio
requests